backend/requirements.txt

uvicorn==0.24.0.post1
SQLAlchemy==1.4.50
python-jose==3.3.0
fastapi==0.104.1